ReplaceIstance initializes a instance and might destruct one
Bug: 1354098
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3838291
Commit-Queue: Daniel Brinkers <brinky@google.com>
Cr-Commit-Position: refs/heads/main@{#1036999}

+++ chrome/browser/chromeos/policy/dlp/dlp_rules_manager_impl.cc
@@ -182,8 +182,10 @@ std::pair<DlpRulesManager::Level, absl::optional<T>> GetMaxJoinRestrictionLevel(
 }
 
 void OnSetDlpFilesPolicy(const ::dlp::SetDlpFilesPolicyResponse response) {
-  if (response.has_error_message()) {
+  if (DlpScopedFileAccessDelegate::HasInstance()) {
     DlpScopedFileAccessDelegate::DeleteInstance();
+  }
+  if (response.has_error_message()) {
     LOG(ERROR) << "Failed to set DLP Files policy and start DLP daemon, error: "
                << response.error_message();
     return;
@@ -207,7 +209,9 @@ void OnSetDlpFilesPolicy(const ::dlp::SetDlpFilesPolicyResponse response) {
 
 DlpRulesManagerImpl::~DlpRulesManagerImpl() {
   DataTransferDlpController::DeleteInstance();
-  DlpScopedFileAccessDelegate::DeleteInstance();
+  if (DlpScopedFileAccessDelegate::HasInstance()) {
+    DlpScopedFileAccessDelegate::DeleteInstance();
+  }
 }
 
 // static
@@ -611,7 +615,7 @@ void DlpRulesManagerImpl::OnPolicyUpdate() {
     DlpBooleanHistogram(dlp::kFilesDaemonStartedUMA, true);
     chromeos::DlpClient::Get()->SetDlpFilesPolicy(
         request_to_daemon, base::BindOnce(&OnSetDlpFilesPolicy));
-  } else {
+  } else if (DlpScopedFileAccessDelegate::HasInstance()) {
     DlpScopedFileAccessDelegate::DeleteInstance();
   }
 }
